# Remove dead CRC32 processor code