… include/contractor/crc32_processor.hpp