# Issue

It's dead, Jim.


<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1099.55<br/>plain u32: 1091.82<br/>aliased double: 969.199<br/>plain double: 953.029 | aliased u32: 1099.98<br/>plain u32: 1097.05<br/>aliased double: 960.155<br/>plain double: 953.464 |
| json-render | String: 8.50726ms<br/>Stringstream: 11.7605ms<br/>Vector: 7.66585ms | String: 8.48368ms<br/>Stringstream: 11.6249ms<br/>Vector: 7.61992ms |
| match_ch | Default radius: <br/>4.49462ms/req at 82 coordinate<br/>0.0548125ms/coordinate<br/>Radius 5m: <br/>4.48093ms/req at 82 coordinate<br/>0.0546455ms/coordinate<br/>Radius 10m: <br/>15.2447ms/req at 82 coordinate<br/>0.185911ms/coordinate<br/>Radius 15m: <br/>37.1495ms/req at 82 coordinate<br/>0.453043ms/coordinate<br/>Radius 30m: <br/>315.572ms/req at 82 coordinate<br/>3.84844ms/coordinate | Default radius: <br/>4.48511ms/req at 82 coordinate<br/>0.0546965ms/coordinate<br/>Radius 5m: <br/>4.4992ms/req at 82 coordinate<br/>0.0548683ms/coordinate<br/>Radius 10m: <br/>15.265ms/req at 82 coordinate<br/>0.186159ms/coordinate<br/>Radius 15m: <br/>37.1078ms/req at 82 coordinate<br/>0.452534ms/coordinate<br/>Radius 30m: <br/>316.422ms/req at 82 coordinate<br/>3.8588ms/coordinate |
| match_mld | Default radius: <br/>2.94529ms/req at 82 coordinate<br/>0.0359182ms/coordinate<br/>Radius 5m: <br/>2.80056ms/req at 82 coordinate<br/>0.0341532ms/coordinate<br/>Radius 10m: <br/>10.1523ms/req at 82 coordinate<br/>0.123809ms/coordinate<br/>Radius 15m: <br/>26.066ms/req at 82 coordinate<br/>0.317878ms/coordinate<br/>Radius 30m: <br/>302.846ms/req at 82 coordinate<br/>3.69324ms/coordinate | Default radius: <br/>2.82128ms/req at 82 coordinate<br/>0.0344059ms/coordinate<br/>Radius 5m: <br/>2.81277ms/req at 82 coordinate<br/>0.0343021ms/coordinate<br/>Radius 10m: <br/>10.1962ms/req at 82 coordinate<br/>0.124343ms/coordinate<br/>Radius 15m: <br/>26.0711ms/req at 82 coordinate<br/>0.31794ms/coordinate<br/>Radius 30m: <br/>301.675ms/req at 82 coordinate<br/>3.67896ms/coordinate |
| packedvector | random write:<br/>std::vector 9866.13 ms<br/>util::packed_vector 82451.3 ms<br/>slowdown: 8.35701<br/>random read:<br/>std::vector 8558.35 ms<br/>util::packed_vector 33280.7 ms<br/>slowdown: 3.88868 | random write:<br/>std::vector 9861.51 ms<br/>util::packed_vector 82509.2 ms<br/>slowdown: 8.36679<br/>random read:<br/>std::vector 8547.6 ms<br/>util::packed_vector 33169.6 ms<br/>slowdown: 3.88057 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>601.099ms<br/>0.601099ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>379.732ms<br/>0.379732ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>763.748ms<br/>0.763748ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>161.631ms<br/>0.161631ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>103.677ms<br/>0.103677ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>145.61ms<br/>0.14561ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>160.993ms<br/>0.160993ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>103.712ms<br/>0.103712ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>145.116ms<br/>0.145116ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>601.33ms<br/>0.60133ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>379.124ms<br/>0.379124ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>769.494ms<br/>0.769494ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>162.604ms<br/>0.162604ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>104.778ms<br/>0.104778ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>150.818ms<br/>0.150818ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>161.677ms<br/>0.161677ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>110.053ms<br/>0.110053ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>146.846ms<br/>0.146846ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>730.291ms<br/>0.730291ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>470.06ms<br/>0.47006ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>959.991ms<br/>0.959991ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>282.372ms<br/>0.282372ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>168.619ms<br/>0.168619ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>312.373ms<br/>0.312373ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>280.843ms<br/>0.280843ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>168.609ms<br/>0.168609ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>307.45ms<br/>0.30745ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>726.26ms<br/>0.72626ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>466.55ms<br/>0.46655ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>950.389ms<br/>0.950389ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>278.756ms<br/>0.278756ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>167.27ms<br/>0.16727ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>305.145ms<br/>0.305145ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>275.995ms<br/>0.275995ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>168.286ms<br/>0.168286ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>310.076ms<br/>0.310076ms/req |
| rtree | 1 result:<br/>206.835ms ->  0.0206835 ms/query<br/>10 results:<br/>242.888ms ->  0.0242888 ms/query | 1 result:<br/>207.891ms ->  0.0207891 ms/query<br/>10 results:<br/>242.602ms ->  0.0242602 ms/query |
<!-- BENCHMARK_RESULTS_END -->